Update index.js
I know it is sync, but for some reason, when I use like 20 files to compile at once, it gets stuck (or when I kill it with `control+c` becomes  zombie or get a segmentation fault), so, I had to change it to the `readFileSync` and that way it works. Besides, it is still fast to compile `scss` to `css`.